+++ src/admin-template/template-column.spec.js
@@ -134,4 +134,31 @@ describe('TemplateColumn', function() {
 
     });
 
+    describe('isStockBasedColumn', function() {
+
+        var columnsSpy;
+
+        beforeEach(function() {
+            templateColumn = new TemplateColumn(templateColumnJson);
+
+            columnsSpy = jasmine.createSpyObj('columns', ['includes']);
+            spyOn(TEMPLATE_COLUMNS, 'getStockBasedColumns').andReturn(columnsSpy);
+        });
+
+        it('should return true if column is stock based', function() {
+            columnsSpy.includes.andReturn(true);
+
+            expect(templateColumn.isStockBasedColumn()).toBe(true);
+            expect(columnsSpy.includes).toHaveBeenCalledWith(templateColumn.name);
+        });
+
+        it('should return false if column is not stock based', function() {
+            columnsSpy.includes.andReturn(false);
+
+            expect(templateColumn.isStockBasedColumn()).toBe(false);
+            expect(columnsSpy.includes).toHaveBeenCalledWith(templateColumn.name);
+        });
+
+    });
+
 });