+++ src/admin-template/template.spec.js
@@ -277,19 +277,27 @@ describe('Template', function() {
 
     describe('changePopulateStockOnHandFromStockCards', function() {
 
-        it('should change stock on hand source to stock cards', function() {
+        it('should change stock based columns sources to stock cards', function() {
             RequisitionColumnSpy.columnDependencies.andReturn([]);
             template = new TemplateDataBuilder().withPopulateStockOnHandFromStockCards().withColumn({
                 name: TEMPLATE_COLUMNS.STOCK_ON_HAND,
                 source: COLUMN_SOURCES.USER_INPUT
+            }).withColumn({
+                name: TEMPLATE_COLUMNS.BEGINNING_BALANCE,
+                source: COLUMN_SOURCES.USER_INPUT
             }).build();
 
             spyOn(template.columnsMap.stockOnHand, 'isStockDisabledColumn').andReturn(false);
+            spyOn(template.columnsMap.stockOnHand, 'isStockBasedColumn').andReturn(true);
+            spyOn(template.columnsMap.beginningBalance, 'isStockDisabledColumn').andReturn(false);
+            spyOn(template.columnsMap.beginningBalance, 'isStockBasedColumn').andReturn(true);
             spyOn(template.columnsMap.column1, 'isStockDisabledColumn').andReturn(false);
+            spyOn(template.columnsMap.column1, 'isStockBasedColumn').andReturn(false);
 
             template.changePopulateStockOnHandFromStockCards();
 
             expect(template.columnsMap.stockOnHand.source).toEqual(COLUMN_SOURCES.STOCK_CARDS);
+            expect(template.columnsMap.beginningBalance.source).toEqual(COLUMN_SOURCES.STOCK_CARDS);
         });
 
         it('should disable stock columns', function() {
@@ -301,8 +309,10 @@ describe('Template', function() {
 
             spyOn(template.columnsMap.stockOnHand, 'isStockDisabledColumn').andReturn(true);
             spyOn(template.columnsMap.stockOnHand, 'disableColumnsAndChangeSource').andReturn(true);
+            spyOn(template.columnsMap.stockOnHand, 'isStockBasedColumn').andReturn(false);
             spyOn(template.columnsMap.column1, 'isStockDisabledColumn').andReturn(true);
             spyOn(template.columnsMap.column1, 'disableColumnsAndChangeSource').andReturn(true);
+            spyOn(template.columnsMap.column1, 'isStockBasedColumn').andReturn(false);
 
             template.changePopulateStockOnHandFromStockCards();
 
@@ -310,15 +320,24 @@ describe('Template', function() {
             expect(template.columnsMap.column1.disableColumnsAndChangeSource).toHaveBeenCalled();
         });
 
-        it('should change stock on hand source to user input', function() {
+        it('should change stock based columns sources to user input', function() {
             RequisitionColumnSpy.columnDependencies.andReturn([]);
             template = new TemplateDataBuilder().withColumn({
                 name: TEMPLATE_COLUMNS.STOCK_ON_HAND,
-                source: COLUMN_SOURCES.STOCK_CARDS
+                source: COLUMN_SOURCES.USER_INPUT
+            }).withColumn({
+                name: TEMPLATE_COLUMNS.BEGINNING_BALANCE,
+                source: COLUMN_SOURCES.USER_INPUT
             }).build();
+
+            spyOn(template.columnsMap.stockOnHand, 'isStockBasedColumn').andReturn(true);
+            spyOn(template.columnsMap.beginningBalance, 'isStockBasedColumn').andReturn(true);
+            spyOn(template.columnsMap.column1, 'isStockBasedColumn').andReturn(false);
+
             template.changePopulateStockOnHandFromStockCards();
 
             expect(template.columnsMap.stockOnHand.source).toEqual(COLUMN_SOURCES.USER_INPUT);
+            expect(template.columnsMap.beginningBalance.source).toEqual(COLUMN_SOURCES.USER_INPUT);
         });
     });